Qrs de prueba.

+++ app/models/qrs/qr.py
@@ -1,9 +1,12 @@
+import datetime
+
 import qrcode
 import os
 
 from app.models.baseModel import BaseModel
 from app.models.reservations.reservation import Reservation
 from config import basedir
+
 
 class QR(BaseModel):
     def __init__(self, _id=None, url=None):
@@ -20,22 +23,21 @@ class QR(BaseModel):
         qr = cls(**{})
         qr.url = f'gokartmania.com/#/reservation/{reservation._id}'
         img = qrcode.make(qr.url)
-        #if not os.path.exists("app/reservation_qrs"):
-        #    os.makedirs('app/reservation_qrs')
-        img_path = f'{basedir}/app/reservation_qrs/{qr._id}.png'
-        #img_file = open(img_path,"w+")
+        date_str = reservation.date.strftime("%Y-%m-%d")
+        img_path = f'{basedir}/app/reservation_qrs/{qr._id}{date_str}.png'
         img.save(img_path, format="PNG")
-        #img_file.close()
-        return qr._id+".png"
+        return qr._id + ".png"
 
     @staticmethod
-    def remove_reservations_qrs():
+    def remove_reservation_qrs():
         folder = 'app/reservation_qrs'
         for file in os.listdir(folder):
             file_path = os.path.join(folder, file)
             try:
                 if os.path.isfile(file_path):
-                    os.unlink(file_path)
-                return os.listdir(folder)
+                    qr_date = datetime.datetime.strptime(file_path[-14:-4], "%Y-%m-%d")
+                    today = datetime.datetime.strptime(datetime.datetime.now().strftime("%Y-%m-%d"), "%Y-%m-%d")
+                    if today > qr_date:
+                        os.unlink(file_path)
             except Exception as e:
                 print(e)
